Update the Android app framework for Flutter app
This is done by
$ rm -rf android/
$ flutter create .

In preparation for adding MethodChannels/platform specific code,
it was recommended to get the latest android app framework from Flutter.

This had a side-effect of switching to Kotlin, which I'm ok with, but
that can probably be avoided if necessary.

We may want to do this for iOS too