Fix scale question scaling
Radio buttons should shrink, if necessary, to fit on one row
Labels should stick to the buttons
Center the whole thing (or use the whole screen)

Fixes #114

Note: This also "fixes" the overlap in 112, although things can get very
tiny with the long labels. Will address in a separate commit.