create_deb_pkg.sh: stopped copying to /usr/lib.
The shared libs required by taqo are either linked via ld or loaded by
Dart_LoadELF. The former will locate the so files in the run path
$ORIGIN/lib, which is a relative path to the executable. The latter will
only search in $ORIGIN/lib in our case. So there is no need to copy so
files into /usr/lib, which may interfere with other flutter apps
installed on the system.